Edit Location
1. 물품 위치 거리 계산 표시 수정
2. 위치 모델 수정

+++ bilrun/lib/model/product_list_model.dart
@@ -42,8 +42,8 @@ class ProductList {
   String price;
   String priceProp;
   User user;
-  int lat;
-  int lng;
+  double lat;
+  double lng;
   String photo1;
   dynamic photo2;
   dynamic photo3;
